Stuff changed to account for reflections
Getting color from a ray is now a method so it can be called recursively
in order to get reflected color. Not sure how to determine which
surfaces should be reflective, but the functionality should be easy to
add once I work that out. Based on a bug I had, it should be able to
find out if it reflects onto an object fine.